% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gencode_txdb.R
\name{gencode_txdb}
\alias{gencode_txdb}
\title{Create a Gencode TxDb object}
\usage{
gencode_txdb(version = "31", genome = "hg38", chrs = paste0("chr",
  c(1:22, "X", "Y", "M")))
}
\arguments{
\item{version}{A \code{character(1)} with the Gencode version number.}

\item{genome}{A \code{character(1)} with the human genome version number. Valid
options are \code{'hg38'} or \code{'hg19'}.}

\item{chrs}{A \code{character()} vector with the chromosome (contig) names to
keep.}
}
\value{
A \link[GenomicFeatures:TxDb-class]{GenomicFeatures::TxDb} object.
}
\description{
This function builds a transcript database (TxDb) object which you can then
use to build a Gencode Genomic State object. This function will download
the data from Gencode, import it into R, process it and build the TxDb
object.
}
\examples{

## Build a TxDb object for Gencode v31 chromosome 22
txdb_v31_hg38_chr22 <- gencode_txdb(chrs = 'chr22')

## Explore the result
txdb_v31_hg38_chr22

}
\references{
Based on code for the \code{brainflowprobes} package at:
\url{https://github.com/LieberInstitute/brainflowprobes/blob/master/data-raw/create_sysdata.R}
}
\seealso{
\link{gencode_annotated_genes} \link{gencode_genomic_state}
}
\author{
Leonardo Collado-Torres
}
