% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenomicState-package.R
\docType{package}
\name{GenomicState-package}
\alias{GenomicState}
\alias{GenomicState-package}
\title{GenomicState: Build and access GenomicState objects for use with derfinder tools from
    sources like Gencode}
\description{
This package contains functions for building GenomicState objects
    from different annotation sources such as Gencode. It also provides access
    to these files at JHPCE.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/LieberInstitute/GenomicState}
  \item Report bugs at \url{https://support.bioconductor.org/t/GenomicState}
}

}
\author{
\strong{Maintainer}: Leonardo Collado-Torres \email{lcolladotor@gmail.com} (\href{https://orcid.org/0000-0003-2140-308X}{ORCID})

}
\keyword{internal}
